% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BIC_IS.R
\name{BIC_IS}
\alias{BIC_IS}
\title{Variable selection for regularized SCA using BIC and Index of Sparseness.}
\usage{
BIC_IS(DATA, Jk, R, NRSTARTS, MaxIter, LassoSequence, GLassoSequence,
  method)
}
\arguments{
\item{DATA}{The concatenated data block, with rows representing subjects.}

\item{Jk}{A vector. Each element of this vector is the number of columns of a data block.}

\item{R}{The number of components (R>=2).}

\item{NRSTARTS}{The number of multistarts for this algorithm. The default value is 5.}

\item{MaxIter}{Maximum number of iterations for this algorithm. The default value is 400.}

\item{LassoSequence}{The range of Lasso tuning parameters. The default value is a sequence of 20 numbers from 0.00000001
to the smallest Lasso tuning parameter value that makes all the component loadings equal to zero. Note that by default the 20 numbers are equally spaced on the log scale.}

\item{GLassoSequence}{The range of Group Lasso tuning parameters. The default value is a sequence of 20 numbers from 0.00000001
to the smallest Group Lasso tuning parameter value that makes all the component loadings equal to zero. Note that by default the 20 numbers are equally spaced (but not on the log scale).}

\item{method}{"datablock" or "component". These are two options with respect to the grouping of the loadings as used in the Group Lasso penalty. 
If \code{method="component"}, the block-grouping of the coefficients is applied per component separately. If \code{method = "datablock"}, the grouping
is applied on the concatenated data block, with loadings of all components together. If \code{method} is missing, then the "component" method is used 
by default.}
}
\value{
\item{BIC}{A matrix of BIC values.}
\item{IS}{A matrix of IS values.}
\item{BIC_tuning}{Recommended tuning parameters for Lasso and Group Lasso based on BIC}
\item{IS_tuning}{Recommended tuning parameters for Lasso and Group Lasso based on IS}
}
\description{
\code{BIC_IS} is used to find the proper conbination of Lasso and Group Lasso tuning parameters for regularized SCA based on BIC and Index of Sparseness.
}
\examples{
\dontrun{
DATA1 <- matrix(rnorm(50), nrow=5)
DATA2 <- matrix(rnorm(100), nrow=5)  
DATA <- cbind(DATA1, DATA2)
Jk <- c(10, 20) 
BIC_IS(DATA, Jk, R=5, NRSTARTS = 40, MaxIter = 100)
}

}
\references{
Gajjar, S., Kulahci, M., & Palazoglu, A. (2017). Selection of non-zero loadings in sparse principal component analysis. \emph{Chemometrics and Intelligent Laboratory Systems}, \emph{162}, 160-171.

Trendafilov, N. T. (2014). From simple structure to sparse components: a review. \emph{Computational Statistics}, \emph{29}(3-4), 431-454.

Zou, H., Hastie, T., & Tibshirani, R. (2006). Sparse principal component analysis. \emph{Journal of computational and graphical statistics}, \emph{15}(2), 265-286.

Croux, C., Filzmoser, P., & Fritz, H. (2013). Robust sparse principal component analysis. \emph{Technometrics}, \emph{55}(2), 202-214.

Guo, J., James, G., Levina, E., Michailidis, G., & Zhu, J. (2010). Principal component analysis with sparse fused loadings. \emph{Journal of Computational and Graphical Statistics}, \emph{19}(4), 930-946.
}
