% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDfriedman.R
\name{CDfriedman}
\alias{CDfriedman}
\title{Variable selection with Lasso and Group Lasso}
\usage{
CDfriedman(DATA, Jk, R, LASSO, GROUPLASSO, MaxIter)
}
\arguments{
\item{DATA}{A matrix, which contains the concatenated data with the same subjects from multiple blocks.}

\item{Jk}{A vector containing number of variables in the concatinated data matrix.}

\item{R}{Number of components.}

\item{LASSO}{A Lasso tuning parameter.}

\item{GROUPLASSO}{A group Lasso tuning parameter.}

\item{MaxIter}{The maximum rounds of iterations. It should be a positive integer. The default value is 400.}
}
\value{
\item{Pmatrix}{Estimated component loading matrix (i.e., P).}
\item{Tmatrix}{Estimated component score matrix (i.e., T).}
\item{Lossvec}{A vector containing the loss in each iteration.}
\item{Loss}{The loss in the final iteration.}
}
\description{
Variable selection with Lasso and Group Lasso penalties to identify component and distinctive components.
}
\examples{
\dontrun{
DATA1 <- matrix(rnorm(50), nrow=5)
DATA2 <- matrix(rnorm(100), nrow=5) #thus, we assume that DATA1 and DATA2 are with respect to the same 5 subjects here.
DATA <- cbind(DATA1, DATA2)
Jk <- c(10, 20) #DATA1 has 10 columns, DATA2 20.
R <- 5 # assume that for some reason, we want to have 5 components in the concatenated P matrix.
LASSO <- 0.2 # assume that we know LASSO=0.2 is a suitable value.
GROUPLASSO <- 0.4 # assume that Group Lasso =0.4 is a suitable value.
MaxIter <- 400
results <- CDfriedman(DATA, Jk, R, LASSO, GROUPLASSO, MaxIter)

results$Pmatrix # to check the final concatenated component loading matrix.
}
}
\references{
Friedman, J., Hastie, T., & Tibshirani, R. (2010). A note on the group lasso and a sparse group lasso. arXiv preprint arXiv:1001.0736.

Yuan, M., & Lin, Y. (2006). Model selection and estimation in regression with grouped variables. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 68(1), 49-67.
}

