% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_structure.R
\name{component_structure}
\alias{component_structure}
\title{target <- matrix(c(1,1,1,0,1,0,0,1,0,1), 2, 5)
Jk <- c(144, 44)
R <- 5
component_structure(Jk, R, target)}
\usage{
component_structure(Jk, R, target)
}
\description{
target <- matrix(c(1,1,1,0,1,0,0,1,0,1), 2, 5)
Jk <- c(144, 44)
R <- 5
component_structure(Jk, R, target)
}
