% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossvali.R
\name{crossvali}
\alias{crossvali}
\title{A cross-validation procedure when common/distictive processes are known.}
\usage{
crossvali(DATA, Jk, R, CommPosition, component_structure, MaxIter, NRSTARTS,
  LassoSequence)
}
\arguments{
\item{DATA}{The concatenated data block, with rows represending subjects.}

\item{Jk}{A vector. Each element of this vector is the number of columns of a data block.}

\item{R}{The number of components.}

\item{CommPosition}{A number (vector) indicating which component(s) is the common component(s).}

\item{component_structure}{A matrix specifing which elements in the component matrix should be fixed at zeros.
see \code{component_structure}.}

\item{MaxIter}{Maximum number of iterations for this algorithm. The default value is 400.}

\item{NRSTARTS}{The number of multistarts for this algorithm. The default value is 20.}

\item{LassoSequence}{The range of lasso tuning parameters. The default value is a sequence of 10 numbers from 0
to the smallest Lasso tuning parameter that can make the entire common component(s) to be zeros.}
}
\value{
\item{PRESS}{A vector of predicted residual sum of squares (PRESS) for the sequence of Lasso tuning parameters.}
\item{LassoSeqence}{The sequence of Lasso tuning parameters used in cross-validation.}
\item{plot}{A plot of PRESS against Lasso tuning parameters. }
}
\description{
\code{crossvali} helps to find a range of lasso tuning parameters for the common component so as to generate sparse common component.
}
\details{
This function search through a range of lasso tuning parameters for the common component, while keeping
distinctive components fixed (- that is, the zeros in the distinctive components are fixed). This function
may be of help if user wants to obtain some sparsness in the common component.
}
\examples{
crossvali(DATA, Jk, R, CommPosition, component_structure, MaxIter = 100, NRSTARTS = 40, LassoSequence = seq(from= 0.002, to=0.1, length.out = 10))
}
\references{
Bro, R., Kjeldahl, K., Smilde, A. K., & Kiers, H. A. L. (2008). Cross-validation of component models: a critical look at current methods. Analytical and bioanalytical chemistry, 390(5), 1241-1251.

Gu, Z., & Van Deun, K. (2016). A variable selection method for simultaneous component based data integration. \emph{Chemometrics and Intelligent Laboratory Systems}, 158, 187-199.
}

