% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cv_CDfriedman.R
\name{cv_CDfriedman}
\alias{cv_CDfriedman}
\title{A K-fold cross-validation procedure when common/distictive processes are unknown.}
\usage{
cv_CDfriedman(DATA, Jk, R, MaxIter, NRSTARTS, LassoSequence, GLassoSequence,
  nfolds)
}
\arguments{
\item{DATA}{The concatenated data block, with rows represending subjects.}

\item{Jk}{A vector. Each element of this vector is the number of columns of a data block.}

\item{R}{The number of components.}

\item{MaxIter}{Maximum number of iterations for this algorithm. The default value is 400.}

\item{NRSTARTS}{The number of multistarts for this algorithm. The default value is 5.}

\item{LassoSequence}{The range of Lasso tuning parameters. The default value is a sequence of 5 numbers from 0.001
to the smallest Lasso tuning parameter that can make all the components to be zeros.}

\item{GLassoSequence}{The range of Group Lasso tuning parameters. The default value is a sequence of 5 numbers from 0.001
to the smallest Group Lasso tuning parameter that can make all the components to be zeros.}

\item{nfolds}{Number of folds. If missing, then 10 fold cross-validation will be performed.}
}
\value{
\item{PRESS}{A matrix of predicted residual sum of squares (PRESS) for the sequences of Lasso and Group Lasso tuning parameters.}
\item{plot}{A plot of PRESS against Lasso and Group Lasso tuning parameters. Note that on the x axis are the index numbers of
Lasso and Group Lasso tuning parameters, and to find their corresponding values, please make use of \code{lasso_index} and \code{Glasso_index}}
\item{plotSE}{A plot of PRESS +/- 1 standard error against Lasso and Group Lasso tuning parameters. Note that on the x axis are the index numbers of
Lasso and Group Lasso tuning parameters, and to find their corresponding values, please make use of \code{lasso_index} and \code{Glasso_index}}
\item{Lasso_values}{The sequence of Lasso tuning parameters used for cross-validation. For example, suppose from the plot we found that the index number
for Lasso is \code{6}, its corresponding Lasso tuning parameter is \code{Lasso_values[6]}.}
\item{Glasso_values}{The sequence of Group Lasso tuning parameters used for cross-validation. For example, suppose from the plot we found that the index number
for Group Lasso is \code{6}, its corresponding Group Lasso tuning parameter is \code{Glasso_values[6]}.}
}
\description{
\code{cv_CDfriedman} helps to find a range of Lasso and Group Lasso tuning parameters for the common component so as to generate sparse common component.
}
\details{
This function search through a range of Lasso and Group Lasso tuning parameters for identifying common and distinctive components
}
\examples{
cv.CDpre(DATA, Jk, R, CommPosition, component_structure, MaxIter = 100, NRSTARTS = 40, LassoSequence = seq(from= 0.002, to=0.1, length.out = 10))
}
\references{
Witten, D.M., Tibshirani, R., & Hastie, T. (2009), A penalized matrix decomposition, with applications to sparse principal components and canonical correlation analysis. \emph{Biostatistics}, \emph{10}(3), 515-534.

Friedman, J., Hastie, T., & Tibshirani, R. (2010). A note on the group lasso and a sparse group lasso. arXiv preprint arXiv:1001.0736.

Yuan, M., & Lin, Y. (2006). Model selection and estimation in regression with grouped variables. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 68(1), 49-67.
}

