% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySTD.R
\name{mySTD}
\alias{mySTD}
\title{Standardize the given data matrix per column, over the rows}
\usage{
mySTD(DATA)
}
\arguments{
\item{DATA}{A data matrix}
}
\value{
a standardized matrix
}
\description{
Standardize the given data matrix per column, over the rows
}
\examples{

mySTD(matrix(1:12), nrow=3, byrow=T))
}

