% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CVsparseSCA.R
\name{plot.CVsparseSCA}
\alias{plot.CVsparseSCA}
\title{A heatmap of the cross-validation result}
\usage{
\method{plot}{CVsparseSCA}(x, ...)
}
\arguments{
\item{x}{A object for plot.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
A heatmap of mean squared prediction errors (MSPE) against Lasso and Group 
Lasso tuning parameters. Note that on the x-axis 
are Group Lasso tuning parameter values, and the on the y-axis 
are the Lasso tuning parameter values.
}
\examples{
\dontrun{
## S3 method for class 'CVsparseSCA'
plot(x)
}

}
