% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.structuredSCA.R
\name{plot.structuredSCA}
\alias{plot.structuredSCA}
\title{Cross-validation plot}
\usage{
\method{plot}{structuredSCA}(x, ...)
}
\arguments{
\item{x}{A object for plot.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
A plot of mean square errors +/- 1 standard error against 
Lasso tuning parameters. The plot is plotted against a log 
scale of lambda if \code{LassoSequence} is not defined by users.
}
\examples{
\dontrun{
## S3 method for class 'structuredSCA'
plot(x)
}

}
