% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CVsparseSCA.R
\name{summary.CVsparseSCA}
\alias{summary.CVsparseSCA}
\title{Display a summary of the results of \code{cv_sparseSCA()}.}
\usage{
\method{summary}{CVsparseSCA}(object, disp, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'CVsparseSCA'.}

\item{disp}{The default is \code{simple}; in this case, the recommended tuning 
parameter values will be displayed. 
If \code{full}, then information is displayed regarding 1) the 
recommended tuning parameter values, 2) the proper region 
for Lasso tuning parameter values, given a Group Lasso tuning 
parameter value, 3) # of variable selected,
4) Predicted residual sum of squares (PRESS), 
5) standard errors for PRESS, 6) Lasso and Group Lasso tuning 
parameter values that have been evaluated.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{cv_sparseSCA()}.
}
\examples{
\dontrun{
## S3 method for class 'CVsparseSCA'
summary(object, disp="full")
}

}
