% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.DISCOsca.R
\name{summary.DISCOsca}
\alias{summary.DISCOsca}
\title{Display a summary of the results of \code{DISCOsca()}.}
\usage{
\method{summary}{DISCOsca}(object, disp, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'DISCOsca'.}

\item{disp}{The default is \code{"simple"}; in this case, the best-fitted common/distinctive 
structure is displayed. 
If \code{"full"}, then information is displayed regarding 1) the best-fitted 
common/distinctive structure, 2) Estimated component score matrix (i.e., T),
3) Estimated component loading matrix (i.e., P), and 4) Proportion of variance 
per component.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{DISCOsca()}.
}
\examples{
\dontrun{
## S3 method for class 'DISCOsca'
summary(object, disp="full")
}

}
