% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sparseSCA.R
\name{summary.sparseSCA}
\alias{summary.sparseSCA}
\title{Display a summary of the results of \code{sparseSCA()}.}
\usage{
\method{summary}{sparseSCA}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'sparseSCA'.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{sparseSCA()}.
}
\examples{
\dontrun{
## S3 method for class 'sparseSCA'
summary(object)
}

}
