% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.undoS.R
\name{summary.undoS}
\alias{summary.undoS}
\title{Display a summary of the results of \code{undoShrinkage()}.}
\usage{
\method{summary}{undoS}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'undoS'.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{undoShrinkage()}.
}
\examples{
\dontrun{
## S3 method for class 'undoS'
summary(object)
}

}
