% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.VAF.R
\name{summary.VAF}
\alias{summary.VAF}
\title{Display a summary of the results of \code{VAF()}.}
\usage{
\method{summary}{VAF}(object, digits, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'VAF'.}

\item{digits}{Specify the digits.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{VAF()}.
}
\examples{
\dontrun{
## S3 method for class 'VAF'
summary(object)
}

}
