% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuckerCoef.R
\name{TuckerCoef}
\alias{TuckerCoef}
\title{Tucker congruence}
\usage{
TuckerCoef(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{A matrix}

\item{MatrixB}{A matrix, which is to be compared to MatrixA}
}
\value{
Return the permutation order, the Tucker coefficient, and the Tucker vector
}
\description{
Calculate Tucker's coefficient of congruence between columns but after accounting for permutational freedom and reflections
}
\examples{
TuckerCoef(X, Y)
}

