% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDfriedman.R
\name{CDfriedman}
\alias{CDfriedman}
\title{Variable selection with Lasso and Group Lasso}
\usage{
CDfriedman(DATA, Jk, R, LASSO, GROUPLASSO, MaxIter)
}
\arguments{
\item{DATA}{A matrix, which contains the concatenated data with the same subjects from multiple blocks.}

\item{Jk}{A vector containing number of variables in the concatinated data matrix.}

\item{R}{Number of components.}

\item{LASSO}{A Lasso tuning parameter.}

\item{GROUPLASSO}{A group Lasso tuning parameter.}

\item{MaxIter}{The maximum rounds of iterations. It should be a positive integer. The default value is 400.}
}
\value{
\item{Pmatrix}{Estimated component loading matrix (i.e., P).}
\item{Tmatrix}{Estimated component score matrix (i.e., T).}
\item{Lossvec}{A vector containing the loss in each iteration.}
}
\description{
Variable selection with Lasso and Group Lasso penalties to identify component and distinctive components.
}
\examples{
CDfriedman(DATA, Jk, R, LASSO, GROUPLASSO, MaxIter)

}
\references{
Friedman, J., Hastie, T., & Tibshirani, R. (2010). A note on the group lasso and a sparse group lasso. arXiv preprint arXiv:1001.0736.

Yuan, M., & Lin, Y. (2006). Model selection and estimation in regression with grouped variables. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 68(1), 49-67.
}

