% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDfriedman.R
\name{CDfriedman}
\alias{CDfriedman}
\title{Variable selection with Lasso and Group Lasso}
\usage{
CDfriedman(DATA, Jk, R, LASSO, GROUPLASSO, MaxIter)
}
\description{
Variable selection with Lasso and Group Lasso
}

