% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpre.R
\name{CDpre}
\alias{CDpre}
\title{Variable selection algorithm with a predefined component loading structure}
\usage{
CDpre(DATA, Jk, R, CommPosition, GroupStructure, LASSO, MaxIter)
}
\description{
Variable selection algorithm with a predefined component loading structure
}
