% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CDpre.R
\name{CDpre}
\alias{CDpre}
\title{Variable selection algorithm with a predefined component loading structure}
\usage{
CDpre(DATA, Jk, R, CommPosition, GroupStructure, LASSO, MaxIter)
}
\arguments{
\item{DATA}{A matrix, which contains the concatenated data with the same subjects from multiple blocks.}

\item{Jk}{A vector containing number of variables in the concatinated data matrix.}

\item{R}{Number of components.}

\item{CommPosition}{A scaler or a vector indicating with column(s) in the component loading matrix (P) is the common component}

\item{GroupStructure}{A matrix indicating the predefined common/distinctive structure in the
component loading matrix (P). This matrix can be generated by the function \code{component_structure}.}

\item{LASSO}{A Lasso tuning parameter for the common component. If Lasso=0, then the common component
will not be sparse. In case user wants to generate a vector of Lasso parameters and performs a cross-validation,
then please use the \code{crossvali} function, which incorporates \code{CDpre} algorithm.}

\item{MaxIter}{The maximum rounds of iterations. It should be a positive integer. The default value is 400.}
}
\value{
\item{Pmatrix}{Estimated component loading matrix (i.e., P).}
\item{Tmatrix}{Estimated component score matrix (i.e., T).}
\item{Lossvec}{A vector containing the loss in each iteration.}
\item{Loss}{The loss in the final iteration.}
}
\description{
Variable selection algorithm when the common/distinctive structure is known a priori.
The common component can also be sparse, which is to be estimated by Lasso.
The distinctive components are not sparse in the sense that the entire variables in a component (belonging to a certain block) are either all zeros or non-zeros.
}
\examples{
CDpre(DATA, Jk, R, CommPosition, GroupStructure, LASSO, MaxIter)

}
\references{
Gu, Z., & Van Deun, K. (2016). A variable selection method for simultaneous component based data integration. \emph{Chemometrics and Intelligent Laboratory Systems}, 158, 187-199.
}

