% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/component_structure.R
\name{component_structure}
\alias{component_structure}
\title{A matrix representing common/distinctive processes.}
\usage{
component_structure(Jk, R, target)
}
\arguments{
\item{Jk}{A vector. Each element of this vector is the number of columns of a data block.}

\item{R}{The number of components.}

\item{target}{A matrix containing 0's and 1's. Its number of columns equals to R, and its number of rows equals to the number of blocks to be integrated. Thus, if the element in
the first row and first column is 1, then it means that the component belonging to the first block and the first component is selected; if it is 0, then the component is fixed at zeros.}
}
\value{
A matrix indicating which cells in the component loading matrix should be fixed at zero.
}
\description{
\code{component_structure} generates a matrix representing the common/distinctive components of component loading matrix.
}
\details{
This function generates a matrix indicating which cells in the component loading matrix should be fixed at zeros.
Thus, it is used together with the function \code{VarSelectComDist}.
}
\examples{
target <- matrix(c(1,1,1,0,1,0,0,1,0,1), 2, 5)
Jk <- c(144, 44)
R <- 5
component_structure(Jk, R, target)
}
\references{
Gu, Z., & Van Deun, K. (2016). A variable selection method for simultaneous component based data integration. \emph{Chemometrics and Intelligent Laboratory Systems}, 158, 187-199.
}
