% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MezzichSolomon.R
\docType{data}
\name{MezzichSolomon}
\alias{MezzichSolomon}
\title{22 experienced psychiatrists' ratings on 17 symptoms for archetypical patients in two psychiatric groups}
\format{The dataset contains the following list:
\describe{
 \item{depr_manicdepre}{A 22x17 matrix of psychiatrists x symptoms for the "depressed_manicdepressed" group.}
 \item{simpleschizo}{A 2x17 matrix of psychiatrists x symptoms for the "schizophrenic" group.}
}}
\source{
The data is originally from Mezzich, J. E., & Solomon, H. (1980). Quantitative studies in social relations. London,
UK: Academic Press.
}
\usage{
MezzichSolomon
}
\description{
This dataset contains 22 experienced psychiatrists' rating on 17 symptoms shown in two psychiatric groups. 
Each psychiatrist rated to what extent the 17 sypmtoms matched certain archetypcal psychiatric patients in 
a grouped classified as "depressed_manicdepressed" and patients in a group classified as "schizophrenic".
}
\note{
Note that in the original study (Mezzich & Solomon, 1980), the data contained 22 psychiatrists' rating on 4 groups 
of 17 symtoms. The data in this package contain 2 of the groups
}
\keyword{datasets}
