% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mySTD.R
\name{mySTD}
\alias{mySTD}
\title{Standardize the given data matrix per column, over the rows. (This function will no longer be exported
and it will eventually removed from the package.)}
\usage{
mySTD(DATA)
}
\arguments{
\item{DATA}{A data matrix}
}
\value{
a standardized matrix
}
\description{
Standardize the given data matrix per column, over the rows. (This function will no longer be exported
and it will eventually removed from the package.)
}
\note{
More details regarding data pre-processing, please see:

Van Deun, K., Smilde, A.K., van der Werf, M.J., Kiers, H.A.L., & Mechelen, I.V. (2009). A structured overview of simultaneous component based data integration. \emph{BMC Bioinformatics}, 10:246.
}
\examples{
\dontrun{
mySTD(matrix(1:12, nrow = 3, ncol = 4))
}
}
