% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gca.R
\name{pca_gca}
\alias{pca_gca}
\title{PCA-GCA method}
\usage{
pca_gca(DATA, Jk, cor_min)
}
\arguments{
\item{DATA}{A concatinated data matrix with the same number of rows.}

\item{Jk}{A vector containing number of variables  in the concatinated data matrix. Please see the example below.}

\item{cor_min}{The minimum corelation bewtween two components. The default value is .9; thus, it means that if the correlation
between the two component is at least .9, then these two components are regarded as forming a single common component.}
}
\value{
Return a matrix indicating the common and distincive components.
}
\description{
Use PCA-GCA method to identify common and distictive components.
}
\examples{
DATA <- c(DATA1, DATA2)
Jk <- c(ncol(DATA1), ncol(DATA1))
pca_gca(DATA, Jk, cor_min = .8)
}

