% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_gca.R
\name{pca_gca}
\alias{pca_gca}
\title{PCA-GCA method}
\usage{
pca_gca(DATA, Jk, cor_min)
}
\arguments{
\item{DATA}{A concatinated data matrix with the same number of rows.}

\item{Jk}{A vector containing number of variables  in the concatinated data matrix. Please see the example below.}

\item{cor_min}{The minimum corelation bewtween two components. The default value is .9; thus, it means that if the correlation
between the two component is at least .9, then these two components are regarded as forming a single common component.}
}
\value{
It prints out the number of components of each block and the number of common components. It also returns the component scores for each block for further analysis.
}
\description{
Use PCA-GCA method to identify common and distictive components.
}
\note{
Please be ware of the interactive input: The function first performs PCA on each data block and then displays the eigenvalues (and a scree plot).
Afterwards the function awaits the input from the user - it needs to know how many components need to be retained for that block.
}
\examples{
DATA <- c(DATA1, DATA2)
Jk <- c(ncol(DATA1), ncol(DATA1))
pca_gca(DATA, Jk, cor_min = .8)

}
\references{
Tenenhaus, A., & Tenenhaus, M. (2011). Regularized generalized canonical correlation analysis. Psychometrika, 76(2), 257-284
}

