% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.CVsparseSCA.R
\name{plot.CVsparseSCA}
\alias{plot.CVsparseSCA}
\title{Ploting Cross-validation results}
\usage{
\method{plot}{CVsparseSCA}(x, ...)
}
\arguments{
\item{x}{A object for plot.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
A plot of PRESS +/- 1 standard error against Lasso OR Group 
Lasso tuning parameters, with the vertical dotted black line 
indicating the lowest PRESS+1SE.
}
\details{
In case both the Lasso sequence and the Group Lasso sequence 
contain more than 2 elements, the cross-validation plot is replaced with 
a heatmap of mean squared prediction errors (MSPE) against Lasso and Group 
Lasso tuning parameters (x-axis: the Group Lasso; y-axis: the Lasso)
}
\examples{
\dontrun{
## S3 method for class 'CVsparseSCA'
plot(x)
}

}
