% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CVsparseSCA.R
\name{summary.CVsparseSCA}
\alias{summary.CVsparseSCA}
\title{Display a summary of the results of \code{cv_sparseSCA()}.}
\usage{
\method{summary}{CVsparseSCA}(object, disp, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'CVsparseSCA'.}

\item{disp}{The default is \code{"tuning"}; in this case, the recommended tuning 
parameter values are presented. 
If \code{"estimatedPT"}, then the estimated component loading and estimated
component score matrices (based on the recommended tuning paramter
values) are presented.
If \code{"full"}, then information is displayed regarding 1) the 
recommended tuning parameter values, 2) the estimated component 
loading and estimated component score matrices (based on the 
recommended tuning paramter values), 3) the proper region 
for Lasso tuning parameter values, given a Group Lasso tuning 
parameter value, 4) # of variable selected,
5) Predicted residual sum of squares (PRESS), 
6) standard errors for PRESS, 7) Lasso and Group Lasso tuning 
parameter values that have been evaluated.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{cv_sparseSCA()}.
}
\examples{
\dontrun{
## S3 method for class 'CVsparseSCA'
summary(object, disp="full")
}

}
