% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.CVstructuredSCA.R
\name{summary.CVstructuredSCA}
\alias{summary.CVstructuredSCA}
\title{Display a summary of the results of \code{cv_structuredSCA()}.}
\usage{
\method{summary}{CVstructuredSCA}(object, disp, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'CVstructuredSCA'.}

\item{disp}{The default is \code{"tuning"}; in this case, the recommended tuning 
parameter values for Lasso is displayed
If \code{"estimatedPT"}, then the estimated component loading and
component score matrices (given the recommended tuning parameter)
is displayed. 
If \code{"full"}, then information is displayed regarding 1) the 
recommended tuning parameter values for Lasso, 2) the estimated component 
loading and component score matrices, 3) the proper region 
for Lasso tuning parameter values, based on the 1SE rule, 4) mean squared
prediction error (MSPE), 5) Lasso tuning 
parameter values that have been evaluated.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{cv_structuredSCA()}.
}
\examples{
\dontrun{
## S3 method for class 'CVstructuredSCA'
summary(object, disp="full")
}

}
