% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.structuredSCA.R
\name{summary.structuredSCA}
\alias{summary.structuredSCA}
\title{Display a summary of the results of \code{structuredSCA()}.}
\usage{
\method{summary}{structuredSCA}(object, ...)
}
\arguments{
\item{object}{Object of class inheriting from 'structuredSCA'.}

\item{...}{Argument to be passed to or from other methods.}
}
\description{
Display a summary of the results of \code{structuredSCA()}.
}
\examples{
\dontrun{
## S3 method for class 'structuredSCA'
summary(object)
}

}
