% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TuckerCoef.R
\name{TuckerCoef}
\alias{TuckerCoef}
\title{Tucker congruence}
\usage{
TuckerCoef(MatrixA, MatrixB)
}
\arguments{
\item{MatrixA}{A matrix}

\item{MatrixB}{A matrix, which is to be compared to MatrixA}
}
\value{
\item{perm}{the permutation order.}
\item{tucker_value}{the Tucker coefficient.}
\item{tucker_vector}{the Tucker vector.}
}
\description{
\code{TuckerCoef} calculate Tucker's coefficient of congruence between columns but after accounting for permutational freedom and reflections
}
\examples{
maxtrix1 <- matrix(rnorm(50), nrow=5)
maxtrix2 <- matrix(rnorm(50), nrow=5)
TuckerCoef(maxtrix1, maxtrix2)
}

